requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# AI Providers
openai==1.3.7
anthropic==0.7.8

# Enhanced Search (Optional)
google-search-results==2.4.2

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1 